Add SfmSketch for private cardinality estimation
The SFM sketch is a mergeable, differentially private sketch
for distinct counting, introduced in
https://arxiv.org/abs/2302.02056. Conceptually, it is similar
to HyperLogLog, except that noise may be intentionally added to
the sketch to preserve the privacy of its entries. After noise
has been added, the sketch becomes private and immutable.
However, private sketches may be merged with other private
sketches to obtain a sketch of their union.

This sketch may be used to obtain private estimates of distinct
counts with low memory footprint in a distributed setting.

Some minor refactoring of the PrivateLpcaSketch is included in
this commit. (The PrivateLpcaSketch can be used to convert an
existing HyperLogLog to a different, less flexible privacy-
preserving sketch. While similar in spirit, the use cases for
these two sketches are subtly different.)