Upgrade to Jetty 12
This commit upgrades the codebase to use jetty 12 along with all
the dependent modules. There are a large number of changes outside
of jetty that are required to comply with the updated APIs as well
as some of the newer modernizer rules that execute with the target
bytecode of some modules set to 17

The main types of changes are outlined below:

1. Nearly all usage of packages from javax.annotation, javax.validation
javax.inject, javax.ws.rs-api, and javax.servlet have been replaced with
the newer jakarta.* APIs. This is a requirement due to the newer jetty
version requiring these libraries in order properly function. Some
annotations have been removed entirely from the jakarta.annotation
package, so they have been replaced with equivalent annotations from
com.google.errorprone:error_prone_annotations package. Guice was
upgraded to 7.0 to support the jakarta.inject namespace

2. Minimum bytecode level on some modules being raised to 17. Due to
Jetty 12 being compiled for Java 17 bytecode, it means modules which rely
on jetty directly and transitively must be compiled with java 17 bytecode
as well. http-client and http-server are the modules with direct
dependencies on jetty, but many other such as event, discovery, and more
have their bytecode level increased in order to accomodate the jetty
upgrade. Bytecode levels are set in the project's pom.xml <properties>
tag using <project.build.targetJdk>

3. Modernizer changes. Due to modules being upgraded to Java 17 bytecode
some new modernizer rules are now enforced to take advantage of the
new Java 17 APIs. Many modules have slight changes to comply with our
modernizer plugin. Most changes center around using Path instead of File
and Java Streams API equivalents in place of Guava Iterables and Functions

4. The configuration's validator implementation is updated to use the
hibernate validator instead of Apache bval-jsr. This is because the
bval-jsr dependency does not support the jakarta namespace annotations
at a java 8 bytecode level. Hibernate is used because it satisfies the
requirement to keep this module at Java 8 bytecode.

Aside from these, the rest of the changes center around supporting Jetty
12's new APIs within our http-client and http-server implementation.
For example, supporting the new ContentProvider APIs and changing usages
of deprecated or removed methods from Jetty 9.