Explicitly set the netty SSL Provider to JDK
This code uses the FATAL_ALERT failure behavior, which is only supported
by the JDK ssl provider. However, if no SSL Provider is set, it is
environment dependent whether the JDK or OpenSSL provider is used.
This change Explicitly sets the ssl provider to JDK.

This fixes an issue where servers might fail on start up in
ReferenceCountedOpenSslContext with "OpenSSL provider does not support
FATAL_ALERT behavior" if they get created with OPEN_SSL as the provider.

Netty code pointers:
https://github.com/netty/netty/blob/4.1/handler/src/main/java/io/netty/handler/ssl/JdkSslContext.java#L420
https://github.com/netty/netty/blob/4.1/handler/src/main/java/io/netty/handler/ssl/ReferenceCountedOpenSslContext.java#L777
https://github.com/netty/netty/blob/4.1/handler/src/main/java/io/netty/handler/ssl/SslContext.java#L832
https://github.com/netty/netty/blob/4.1/handler/src/main/java/io/netty/handler/ssl/SslContext.java#L127-L135